requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15,<3.16
psycopg2-binary>=2.9,<3.0
drf-spectacular>=0.27,<0.28
python-decouple>=3.8,<4.0
